Update student testimonial section in index.html
This commit corrects the course level of a student testimonial in index.html, changing it from "A/L Edexcel Biology" to "AS Edexcel Biology". This change is necessary to ensure the accuracy of the information presented on the website, hence maintaining the trust of website users. This information primarily pertains to prospective students who are comparing course levels and past student experiences.