Add gitignore and update index.html for Studytomy-Firebase project
This commit includes adding a .gitignore file for the Studytomy-Firebase project in the .idea directory to ignore unnecessary files from version control. It also contains 'encodings.xml', 'indexLayout.xml', and 'vcs.xml' for better project settings. Additionally, an external script in the 'index.html' has been removed and replaced the image source of a testimonial, along with adding two new testimonial sections.

The external script from 'index.html' might be causing a slowdown on page load, so it's removed. The image was updated to represent the testimonial author appropriately, and new testimonials were added to provide user feedback about the platform.