Remove inline CSS from index.html
The inline CSS present in the index.html file has been removed to keep the HTML file clean and to improve code maintenance. All the styles have been migrated to the external style sheet. This way, changes to the style can be made without altering the HTML structure, making it cleaner and more focused.