Repositioned discount section and updated testimonial details
The 'Get discounted price' and 'Get Referral Discounts' sections were moved for improved visual flow and to enhance user navigation. These sections are strategically placed in the middle of the page to grab the attention of potential clients and provide clarity on the discounts available for group classes and referral program.

Additionally, the locations were included in the testimonial section to provide better context on the source of the testimonials and present our global appeal.