Update website metadata and add responsive styling
This update modifies the website metadata in 'home-school.html' and 'exam-boards.html'. The website title for 'home-school.html' is updated, and in 'exam-boards.html' the description and keywords are adjusted for better SEO performance. Also, a new responsive styling is added to improve the mobile view of the site. Changes also include improving the HTML organisation and updating the sections for various school boards in 'exam-boards.html'.