Update menu links and remove unused id
The relative paths have been corrected in the dropdown links of the exam board. It ensures the user can navigate directly to the particular section on the exam boards page from the dropdown menu. An extraneous id ("apply-aprt") has been removed from a section element, as it wasn't used anywhere in the script or styles.