Update media query max-width in bootstrap CSS
The max-width value in a certain media query within bootstrap.min.css has been updated. It was previously set to 1200px, which is generally meant for larger screens. It was decreased to 986px in order to cater to smaller screen sizes, improving the responsiveness and adaptability of the website design.